+++ example/lib/main-localized.dart
@@ -44,6 +44,7 @@ class Demo extends StatelessWidget {
         const Locale('km', ''), // Khmer, no country code
         const Locale('ko', ''), // Korean, no country code
         const Locale('lt', ''), // Lithuanian, no country code
+        const Locale('mn', ''), // Mongolian, no country code
         const Locale('nb', ''), // Norwegian, no country code
         const Locale('nl', ''), // Dutch, no country code
         const Locale('pt', ''), // Portuguese, no country code
@@ -139,6 +140,7 @@ class DemoLocalizationsDelegate
         'km',
         'ko',
         'lt',
+        'mn',
         'nb',
         'nl',
         'pt',